all channel creation stuff moved into channel manager
